Refine Streamlit dashboard metrics and rain visuals
